# api/v1/types.gen.go
// Package v1 provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.12.4 DO NOT EDIT.
package v1

import (
	"time"
)

// Error defines model for Error.
type Error struct {
	Code    int32  `json:"code"`
	Message string `json:"message"`
}

// NewRole defines model for NewRole.
type NewRole struct {
	CreatedAt   *time.Time    `json:"createdAt,omitempty"`
	Description *string       `json:"description,omitempty"`
	Name        string        `json:"name"`
	Permissions *[]Permission `json:"permissions,omitempty"`
	UpdatedAt   *time.Time    `json:"updatedAt,omitempty"`
}

// Permission defines model for Permission.
type Permission struct {
	Description *string `json:"description,omitempty"`
	Target      string  `json:"target"`
}

// Role defines model for Role.
type Role struct {
	CreatedAt   *time.Time    `json:"createdAt,omitempty"`
	Description *string       `json:"description,omitempty"`
	Id          string        `json:"id"`
	Name        string        `json:"name"`
	Permissions *[]Permission `json:"permissions,omitempty"`
	UpdatedAt   *time.Time    `json:"updatedAt,omitempty"`
}

// GetAssignmentsParams defines parameters for GetAssignments.
type GetAssignmentsParams struct {
	// Subject subject to return assignments for
	Subject string `form:"subject" json:"subject"`

	// Scope scope to return assignments for
	Scope string `form:"scope" json:"scope"`
}

// RemoveRoleAssignmentJSONBody defines parameters for RemoveRoleAssignment.
type RemoveRoleAssignmentJSONBody struct {
	Scope   string `json:"scope"`
	Subject string `json:"subject"`
}

// AssignRoleJSONBody defines parameters for AssignRole.
type AssignRoleJSONBody struct {
	Scope   string `json:"scope"`
	Subject string `json:"subject"`
}

// CreateRoleJSONRequestBody defines body for CreateRole for application/json ContentType.
type CreateRoleJSONRequestBody = NewRole

// UpdateRoleJSONRequestBody defines body for UpdateRole for application/json ContentType.
type UpdateRoleJSONRequestBody = Role

// RemoveRoleAssignmentJSONRequestBody defines body for RemoveRoleAssignment for application/json ContentType.
type RemoveRoleAssignmentJSONRequestBody RemoveRoleAssignmentJSONBody

// AssignRoleJSONRequestBody defines body for AssignRole for application/json ContentType.
type AssignRoleJSONRequestBody AssignRoleJSONBody

// RemoveRolePermissionJSONRequestBody defines body for RemoveRolePermission for application/json ContentType.
type RemoveRolePermissionJSONRequestBody = Permission

// AddRolePermissionJSONRequestBody defines body for AddRolePermission for application/json ContentType.
type AddRolePermissionJSONRequestBody = Permission